+++ aospy/utils/times.py
@@ -335,18 +335,18 @@ def ensure_time_avg_has_cf_metadata(ds):
     RAW_END_DATE_STR = internal_names.RAW_END_DATE_STR
     TIME_BOUNDS_STR = internal_names.TIME_BOUNDS_STR
     TIME_STR = internal_names.TIME_STR
-    NV_STR = internal_names.NV_STR
+    BOUNDS_STR = internal_names.BOUNDS_STR
     TIME_WEIGHTS_STR = internal_names.TIME_WEIGHTS_STR
 
     if TIME_WEIGHTS_STR not in ds:
-        time_weights = ds[TIME_BOUNDS_STR].diff(NV_STR)
+        time_weights = ds[TIME_BOUNDS_STR].diff(BOUNDS_STR)
         time_weights = time_weights.rename(TIME_WEIGHTS_STR).squeeze()
-        ds[TIME_WEIGHTS_STR] = time_weights.drop(NV_STR)
+        ds[TIME_WEIGHTS_STR] = time_weights.drop(BOUNDS_STR)
 
-    avg_start_date = ds[TIME_BOUNDS_STR].isel(**{TIME_STR: 0, NV_STR: 0})
-    ds[RAW_START_DATE_STR] = avg_start_date.drop([TIME_STR, NV_STR])
-    avg_end_date = ds[TIME_BOUNDS_STR].isel(**{TIME_STR: -1, NV_STR: 1})
-    ds[RAW_END_DATE_STR] = avg_end_date.drop([TIME_STR, NV_STR])
+    avg_start_date = ds[TIME_BOUNDS_STR].isel(**{TIME_STR: 0, BOUNDS_STR: 0})
+    ds[RAW_START_DATE_STR] = avg_start_date.drop([TIME_STR, BOUNDS_STR])
+    avg_end_date = ds[TIME_BOUNDS_STR].isel(**{TIME_STR: -1, BOUNDS_STR: 1})
+    ds[RAW_END_DATE_STR] = avg_end_date.drop([TIME_STR, BOUNDS_STR])
 
     for coord in [TIME_BOUNDS_STR, RAW_START_DATE_STR, RAW_END_DATE_STR]:
         ds[coord].attrs['units'] = ds[TIME_STR].attrs['units']